+++ src/RobotStudio.Scripting/IGCodeCommandMapper.cs
@@ -1,0 +1,8 @@
+namespace RobotStudio.Scripting;
+
+public interface IGCodeCommandMapper
+{
+    RobotScriptCompilation Map(
+        GCodeProgram program,
+        RobotScriptParseContext? context = null);
+}