+++ src/RobotStudio.Scripting/RobotScriptUnit.cs
@@ -1,0 +1,6 @@
+namespace RobotStudio.Scripting;
+
+public enum RobotScriptUnit
+{
+    Millimeters
+}